pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relayai"
version = "1.0.0"
description = "Claude thinks. Gemini speaks. Save up to 80% on AI API costs."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [{ name = "RelayAI" }]
keywords = ["ai", "claude", "gemini", "cli", "code", "terminal"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Environment :: Console",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "click>=8.0",
  "rich>=13.0",
  "anthropic>=0.25.0",
  "google-generativeai>=0.5.0",
  "google-auth-oauthlib>=1.0",
  "google-auth>=2.0",
  "keyring>=24.0",
  "requests>=2.28",
]

[project.optional-dependencies]
dev = ["build", "twine", "pytest"]

[project.urls]
Homepage = "https://github.com/yourusername/relayai"
Repository = "https://github.com/yourusername/relayai"
Issues = "https://github.com/yourusername/relayai/issues"

[project.scripts]
relayai = "relayai.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["relayai*"]
